Add image on cloudinary and local

+++ src/Rookie.Application/Products/Commands/DeleteProductCommand/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 using MediatR;
+using Rookie.Application.Contracts.Infrastructure;
 using Rookie.Domain.Common;
 using Rookie.Domain.DomainError;
 using Rookie.Domain.ProductEntity;
@@ -8,9 +9,11 @@ namespace Rookie.Application.Products.Commands.DeleteProductCommand
     public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Result<int>>
     {
         private readonly IProductRepository _productRepository;
-        public DeleteProductCommandHandler(IProductRepository productRepository)
+        private readonly IImageService _imageService;
+        public DeleteProductCommandHandler(IProductRepository productRepository, IImageService imageService)
         {
             _productRepository = productRepository;
+            _imageService = imageService;
         }
         public async Task<Result<int>> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
@@ -26,9 +29,18 @@ namespace Rookie.Application.Products.Commands.DeleteProductCommand
             if (ProductDeleted == null)
                 return Result.Failure<int>(ProductErrors.NotFindProduct);
 
-            _productRepository.Delete(ProductDeleted);
+            //delete on cloud
+            int checkDelete = await _imageService.DeletePhoto("pyfx6fidzml09degawh8");
 
-            return 1;
+            //delete on cloud successfully
+            if (checkDelete == 1)
+            {
+                //delete on local
+                _productRepository.Delete(ProductDeleted);
+                return 1;
+            }
+            else
+                return 0;
         }
     }
 }
